Add CI. Add linter (#16)
---------

+++ syncer/helper/helper.go
@@ -31,8 +31,8 @@ func CreateJobs(txCount, workerCount uint64) []txworker.TxJob {
 		}
 
 		ranges = append(ranges, txworker.TxJob{
-			From: uint64(cursor),
-			To:   uint64(cursor) + size,
+			From: cursor,
+			To:   cursor + size,
 		})
 
 		cursor += size
@@ -90,13 +90,16 @@ func parseUint256Data(data string) (*big.Int, error) {
 	if s == "" || s == "0x" {
 		return big.NewInt(0), nil
 	}
+
 	b, err := hex.DecodeString(strings.TrimPrefix(s, "0x"))
 	if err != nil {
 		return nil, err
 	}
+
 	if len(b) > 32 {
 		return nil, fmt.Errorf("data too long")
 	}
+
 	return new(big.Int).SetBytes(b), nil
 }
